+++ seckill/src/main/java/com/ling/seckill/mapper/GoodsMapper.java
@@ -27,5 +27,5 @@ public interface GoodsMapper extends BaseMapper<Goods> {
      * @param goodsId
      * @return
      */
-    GoodsVo findGoodsVoByGoodsId(Integer goodsId);
+    GoodsVo findGoodsVoByGoodsId(Long goodsId);
 }
