+++ seckill/src/main/java/com/ling/seckill/service/IGoodsService.java
@@ -28,5 +28,5 @@ public interface IGoodsService extends IService<Goods> {
      * @param goodsId
      * @return
      */
-    GoodsVo findGoodsVoByGoodsId(Integer goodsId);
+    GoodsVo findGoodsVoByGoodsId(Long goodsId);
 }
